refactor version control folder structure
