+++ src/features/files/shared/urlSegments.ts
