--- src/features/versionControl/index.ts
@@ -1,7 +1,0 @@
-export * from "./fileMetrics"
-export * from "./files"
-export * from "./ignore"
-export * from "./storage"
-export * from "./track"
-export * from "./urlSegments"
-export * from "./FileMetricsTracker"
