+++ src/pages/content/pages/Folder.tsx
@@ -8,7 +8,7 @@ import WideLayout from "@src/components/layouts/WideLayout"
 import { QuickLinks } from "@src/features/quickAccess/"
 import Table from "@src/features/table/Table"
 import { ThemeToggle } from "@src/components/ui/theme-toggle"
-import { FileMetricsTracker } from "@src/features/versionControl"
+import { FileMetricsTracker } from "@src/features/files"
 import { FolderContent } from "@src/features/parser"
 import { useContext } from "react"
 import { PageStateContext } from "@src/features/router/PageStateContext"
