+++ src/pages/popup/PopupFallback.tsx
@@ -9,7 +9,7 @@ import {
     BASE_URL,
     MMS_BASE_URL,
     MY_SAINT_BASE_URL,
-} from "@src/features/versionControl"
+} from "@src/features/files"
 import { useQuery } from "@tanstack/react-query"
 
 export default function PopupFallback() {
